pyproject.toml
[tool.poetry]
name = "db-lesson-18-4"
version = "0.1.0"
description = ""
authors = ["Evgeny Safonov <jeka-luda@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
